+++ src/main/java/org/example/oddventure/domain/admin/service/AdminMatchService.java
@@ -3,16 +3,9 @@ package org.example.oddventure.domain.admin.service;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.example.oddventure.domain.admin.dto.request.MatchCreateRequest;
-import org.example.oddventure.domain.admin.dto.request.MatchUpdateRequest;
-import org.example.oddventure.domain.admin.dto.response.MatchAdminResponse;
-import org.example.oddventure.domain.admin.exception.AdminErrorCode;
-import org.example.oddventure.domain.admin.exception.AdminException;
 import org.example.oddventure.domain.grid.dto.MatchResultDto;
 import org.example.oddventure.domain.grid.dto.response.MatchFetchResponse;
 import org.example.oddventure.domain.grid.service.GridService;
-import org.example.oddventure.domain.match.entity.Match;
-import org.example.oddventure.domain.match.repository.MatchRepository;
 import org.example.oddventure.domain.match.service.MatchService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -22,57 +15,15 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class AdminMatchService {
 
-    private final MatchRepository matchRepository;
     private final GridService gridService;
     private final MatchService matchService;
-
-    // 매치 생성
-    @Transactional
-    public MatchAdminResponse createMatch(MatchCreateRequest request) {
-        Match match = Match.builder()
-                .matchName(request.matchName())
-                .teamA(request.teamA())
-                .teamB(request.teamB())
-                .startTime(request.startTime())
-                .build();
-        Match savedMatch = matchRepository.save(match);
-
-        return MatchAdminResponse.from(savedMatch);
-    }
-
-    // 매치 정보 수정
-    @Transactional
-    public MatchAdminResponse updateMatch(Long matchId, MatchUpdateRequest request) {
-        Match match = matchRepository.findById(matchId)
-                .orElseThrow(() -> new AdminException(AdminErrorCode.MATCH_NOT_FOUND));
-
-        match.update(
-                request.matchName(),
-                request.teamA(),
-                request.teamB(),
-                request.startTime(),
-                request.status()
-        );
-
-        return MatchAdminResponse.from(match);
-    }
 
     // 매치 일정 연동
     @Transactional
     public void fetchMatches() {
         List<MatchFetchResponse> fetchResponses = gridService.fetchMatches();
 
-        fetchResponses.stream()
-                .filter(dto -> !matchRepository.existsByFetchId(dto.fetchId()))
-                .filter(dto -> !dto.teamA().contains("TBD")) // 미정된 경기
-                .filter(dto -> !dto.teamB().contains("TBD"))
-                .map(dto -> Match.builder()
-                        .fetchId(dto.fetchId())
-                        .matchName(dto.matchName())
-                        .teamA(dto.teamA())
-                        .teamB(dto.teamB())
-                        .startTime(dto.startTime())
-                        .build()).forEach(matchRepository::save);
+        fetchResponses.forEach(matchService::fetchMatches);
     }
 
     // 매치 결과 연동
